requirements.txt
streamlit
boto3
awscli
langchain-community
langchain-huggingface
langchain_aws